+++ slack_topics/__main__.py
@@ -1,0 +1,2 @@
+import slack_topics.bot
+slack_topics.bot.main()