+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.7.2 - 2017-10-17
+- Added `Trending on GitHub` and `Random HackerRank Challenge` to topics.
+- Rewrite `Random Programming Language` with `beautifulsoup4` rather than
+  the `wikipedia` module.
+- Added `url_to_soup` utility.
+- Rename `slack_topics.py` to `bot.py`.
+- Fix some error messages referring to `tb2k`.
+
 ## 0.7.1 - 2017-04-24
 - Fixed packaging. (`SyntaxError` in `setup.py` entry points)